Rename task provider core folder
