+++ src/Plugins/Polyrific.Catapult.TaskProvider.Core/TaskProvider.cs
@@ -10,7 +10,7 @@ using Microsoft.Extensions.Logging;
 using Newtonsoft.Json;
 using Polyrific.Catapult.Shared.Dto.Constants;
 
-namespace Polyrific.Catapult.Plugins.Core
+namespace Polyrific.Catapult.TaskProvider.Core
 {
     public abstract class TaskProvider
     {
